requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
duckduckgo_search==5.3.1b1
pymupdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub
Streamlit
langchain_core
langchain-qdrant
langchain-ollama
unstructured[pdf]
onnx==1.16.1
qdrant-client
langchain-experimental 
neo4j
json-repair
agentops==0.2.6
git+https://github.com/AgentOps-AI/crewAI.git@main
crewai[tools]
torch
transformers 
fastapi
uvicorn
PyPDF2
jinja2
ctransformers
python-multipart
aiofiles
document
python-docx
fpdf
phidata
arxiv
scrapegraphai
playwright
yfinance
packaging
duckduckgo-search
groq
sqlalchemy
pgvector
psycopg[binary]
google-generativeai
streamlit-pdf-viewer
requests
pytz
typing-extensions
black
python-dateutil